moved createPost content into Feed. visuals complete, no functionality
